requirements.txt
h5py==2.7.1
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
matplotlib==2.1.0
nltk==3.2.5
nose==1.3.7
numpy==1.13.3
opencv-python==3.3.0.10
pandas==0.21.0
Pillow==4.3.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
tensorflow-gpu==1.0.1
torch==0.3.1
torchvision==0.2.0

